face-recognition-service/python-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
insightface==0.7.3
onnxruntime==1.16.3
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
python-multipart==0.0.6
supabase==2.0.3
